Adjusted Device Object type decoding to use uint16_t rather than int.  Corrected unit tests.

+++ trunk/bacnet-stack/src/arf.c
@@ -95,7 +95,7 @@ int arf_decode_service_request(
     int tag_len = 0;
     uint8_t tag_number = 0;
     uint32_t len_value_type = 0;
-    int type = 0;       /* for decoding */
+    uint16_t type = 0;       /* for decoding */
 
     /* check for value pointers */
     if (apdu_len && data) {
